+++ tests/testthat/test-ARMApolyroots.R
@@ -3,7 +3,7 @@ test_that("ARMApolyroots works", {
   set.seed(44322)
   inv_roots <- .sample_inv_roots(2)
   inv_roots <- c(inv_roots, Conj(inv_roots))
-  ar_coef <- list(ar = Re(.roots2poly(inv_roots)))
+  ar_coef <- list(ar = Re(.roots2coef(inv_roots)))
   res <- 1 / ARMApolyroots(ar_coef)
 
   expect_equal(Re(res[order(Re(res))]), Re(inv_roots[order(Re(inv_roots))]))