+++ NEWS.md
@@ -1,3 +1,5 @@
+# arima2 (development version)
+
 # arima2 3.1.0
 
 * Fixed bug with plotting models with only MA or only AR coefficients.